+++ SEProject/model/ClassSkillsMatch.java
@@ -1,4 +1,4 @@
-package Core;
+package model;
 
 import java.util.ArrayList;
 import java.util.Arrays;
