+++ SEProject/model/Permission.java
@@ -1,4 +1,4 @@
-package Core;
+package model;
 
 public enum Permission {
