+++ SEProject/model/Skill.java
@@ -1,4 +1,4 @@
-package Core;
+package model;
 
 public enum Skill {
 	ART,
